优化

+++ base/src/main/java/com/pine/base/component/uploader/ui/UploadLinearLayout.java
@@ -44,11 +44,11 @@ public abstract class UploadLinearLayout extends LinearLayout {
         return mHelper.getMaxFileCount();
     }
 
-    protected void setMaxFileSize(long maxFileSize) {
+    protected void setMaxFileSize(int maxFileSize) {
         mHelper.setMaxFileSize(maxFileSize);
     }
 
-    public long getMaxFileSize() {
+    public int getMaxFileSize() {
         return mHelper.getMaxFileSize();
     }
